% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearAlgebra-diag_of_vech.R
\name{diag_of_vech}
\alias{diag_of_vech}
\title{Diagonals of A from vech(A)}
\usage{
diag_of_vech(x, loc = FALSE)
}
\arguments{
\item{x}{Vector.}

\item{loc}{Logical.
If \code{loc = TRUE}, return the location of the diagonal elements
in the input vector.
If \code{loc = FALSE}, return the values of the diagonal elements.}
}
\value{
A vector.
}
\description{
Diagonals of a matrix from its half-vectorization.
}
\details{
Generates a vector of length \eqn{m}
of diagonal elements or location in the input vector
of an \eqn{m \times m} symmetric matrix.
}
\examples{
A <- matrix(
  data = c(
    1.0, 0.5, 0.4,
    0.5, 1.0, 0.6,
    0.4, 0.6, 1.0
  ),
  ncol = 3
)
vechA <- c(1.0, 0.5, 0.4, 1.0, 0.6, 1.0)

diag_of_vech(vechA, loc = FALSE)
diag_of_vech(vechA, loc = TRUE)
}
\references{
\href{https://en.wikipedia.org/wiki/Vectorization_(mathematics)#Half-vectorization}{Wikipedia: Half-vectorization}
}
\seealso{
Other Linear Algebra Functions: 
\code{\link{dcap}()},
\code{\link{kcap}()},
\code{\link{sym_of_vechs}()},
\code{\link{sym_of_vech}()},
\code{\link{vechnames}()},
\code{\link{vechsnames}()},
\code{\link{vechs}()},
\code{\link{vech}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Linear Algebra Functions}
\keyword{linearAlgebra}
\keyword{symmetric}
