% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-hess_l_mvn_generic.R
\name{hess_l_mvn_generic}
\alias{hess_l_mvn_generic}
\title{Hessian Matrix of the Multivariate Normal Distribution - Generic}
\usage{
hess_l_mvn_generic(x, mu, sigmacap)
}
\arguments{
\item{x}{Numeric vector of length \code{k}.
The ith vector of observations.}

\item{mu}{Numeric vector.
Parameter.
Mean vector
\eqn{\boldsymbol{\mu}}.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}
}
\value{
A matrix.
}
\description{
Calculates hessian matrix of the log of the likelihood function
of the multivariate normal distribution
for the ith observation.
}
\examples{
n <- 5
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

xcap <- as.data.frame(
  t(
    rmvn_chol(
      n = n,
      mu = mu,
      sigmacap = sigmacap
    )
  )
)

lapply(
  X = xcap,
  FUN = hess_l_mvn_generic,
  mu = mu,
  sigmacap = sigmacap
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn_generic}()},
\code{\link{l_mvn}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_mvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()},
\code{\link{theta_helper}()}
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
