% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-package.R
\docType{package}
\name{multiNorm}
\alias{multiNorm}
\title{multiNorm: Multivariate Normal Distribution}
\description{
A collection of functions
related to the multivariate normal distribution.
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{multiNorm}
\keyword{package}
