% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmeans_mvn_chol.R
\name{rmeans_mvn_chol}
\alias{rmeans_mvn_chol}
\title{Generate Sample Mean Vector
from the Multivariate Normal Distribution
Using the Cholesky Decomposition}
\usage{
rmeans_mvn_chol(rcap, mu, sigmacap, n, varnames = NULL, list = FALSE)
}
\arguments{
\item{rcap}{Positive integer.
\code{R} variates.}

\item{mu}{Numeric vector.
Parameter.
Mean vector
\eqn{\boldsymbol{\mu}}.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{n}{Positive integer.
Sample size.}

\item{varnames}{Character vector
Optional variable names.}

\item{list}{Logical.
If \code{list = TRUE},
returns a list where each element is a vector of means.
If \code{list = FALSE},
returns a matrix where each row is a vector of means.}
}
\value{
A list (\code{list = TRUE}) or a matrix (\code{list = FALSE}).
}
\description{
Generate Sample Mean Vector
from the Multivariate Normal Distribution
Using the Cholesky Decomposition
}
\section{Dependencies}{
\code{\link[=rmvn_chol]{rmvn_chol()}}
}

\examples{
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

rmeans_mvn_chol(
  rcap = 5,
  mu = mu,
  sigmacap = sigmacap,
  n = 100
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()},
\code{\link{theta_helper}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
