% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmeans_mvn_chol_of_theta.R
\name{rmeans_mvn_chol_of_theta}
\alias{rmeans_mvn_chol_of_theta}
\title{Generate Sample Mean Vector
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Parameter Vector theta Input}
\usage{
rmeans_mvn_chol_of_theta(rcap, x, n, varnames = NULL, list = FALSE)
}
\arguments{
\item{rcap}{Positive integer.
\code{R} variates.}

\item{x}{Numeric vector.
Vector of parameters
\eqn{
    \boldsymbol{\theta}
    =
    \{
      \boldsymbol{\mu}
      \mathrm{vech} \left( \boldsymbol{\Sigma} \right)
    \}^{\prime}
  }.}

\item{n}{Positive integer.
Sample size.}

\item{varnames}{Character vector
Optional variable names.}

\item{list}{Logical.
If \code{list = TRUE},
returns a list where each element is a vector of means.
If \code{list = FALSE},
returns a matrix where each row is a vector of means.}
}
\value{
A list (\code{list = TRUE}) or a matrix (\code{list = FALSE}).
}
\description{
Generate Sample Mean Vector
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Parameter Vector theta Input
}
\section{Dependencies}{
\code{\link[=rmeans_mvn_chol]{rmeans_mvn_chol()}}
\code{\link[=rmvn_chol]{rmvn_chol()}}
}

\examples{
x <- rmeans_mvn_chol_of_theta(
  rcap = 100,
  n = 100,
  x = c(0, 0, 1, 0.5, 1)
)

colMeans(x)
cov(x)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
