% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmvn_chol_of_rhocap.R
\name{rmvn_chol_of_rhocap}
\alias{rmvn_chol_of_rhocap}
\title{Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Correlation Matrix Input}
\usage{
rmvn_chol_of_rhocap(n, x, varnames = NULL, data_frame = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{x}{Numeric matrix.
Parameter.
Correlation matrix
\eqn{\mathbf{P}}.}

\item{varnames}{Character vector
Optional variable names.}

\item{data_frame}{Logical.
If \code{data_frame = TRUE},
returns a \code{data.frame}.
If \code{data_frame = FALSE},
returns a \code{matrix}.}
}
\value{
A matrix (\code{data_frame = FALSE}) or data.frame (\code{data_frame = TRUE}).
}
\description{
Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Correlation Matrix Input
}
\examples{
rhocap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

rmvn_chol_of_rhocap(
  n = 5,
  x = rhocap
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn_generic}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn_generic}()},
\code{\link{l_mvn}()},
\code{\link{mvn_theta_helper}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_mvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
