% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmvn_chol_of_theta.R
\name{rmvn_chol_of_theta}
\alias{rmvn_chol_of_theta}
\title{Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Parameter Vector theta Input}
\usage{
rmvn_chol_of_theta(n, x, varnames = NULL, data_frame = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{x}{Numeric vector.
Vector of parameters
\eqn{
    \boldsymbol{\theta}
    =
    \{
      \boldsymbol{\mu} ,
      \mathrm{vech} \left( \boldsymbol{\Sigma} \right)
    \}^{\prime}
  }.}

\item{varnames}{Character vector
Optional variable names.}

\item{data_frame}{Logical.
If \code{data_frame = TRUE},
returns a \code{data.frame}.
If \code{data_frame = FALSE},
returns a \code{matrix}.}
}
\value{
A matrix (\code{data_frame = FALSE}) or data.frame (\code{data_frame = TRUE}).
}
\description{
Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Parameter Vector theta Input
}
\examples{
rmvn_chol_of_theta(
  n = 5,
  x = c(0, 0, 1, 0.5, 1)
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn_generic}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn_generic}()},
\code{\link{l_mvn}()},
\code{\link{mvn_theta_helper}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_mvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
\keyword{mvn}
\keyword{random}
