% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rvcov_mvn_chol.R
\name{rvcov_mvn_chol}
\alias{rvcov_mvn_chol}
\title{Generate Sample Variance-Covariance Matrix
from the Multivariate Normal Distribution}
\usage{
rvcov_mvn_chol(rcap, sigmacap, gammacap, n, list = TRUE, vec = TRUE)
}
\arguments{
\item{rcap}{Positive integer.
\code{R} variates.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{gammacap}{Numeric matrix.
Parameter.
Asymptotic covariance matrix of the covariance matrix
\eqn{\boldsymbol{\Gamma}}.}

\item{n}{Positive integer.
Sample size.}

\item{list}{Logical.
If \code{list = TRUE},
returns a list where each element is a covariance matrix.
If \code{list = FALSE},
returns a matrix where each row corresponds to
the vectorization of the output matrix (\code{vec = TRUE}) or
the half-vectorization of the output matrix (\code{vec = FALSE})}

\item{vec}{Logical.
This is only evaluated when \code{list = FALSE}.
If \code{vec = TRUE},
returns the vectorization of the covariance matrix
for each \code{R} variate.
If \code{vec = FALSE},
returns the half-covariance matrix
for each \code{R} variate.}
}
\value{
A list (\code{vec = FALSE}) or matrix (\code{vec = TRUE}).
}
\description{
Generate Sample Variance-Covariance Matrix
from the Multivariate Normal Distribution
}
\examples{
sigmacap <- matrix(
  data = c(1, 0.5, 0.5, 1),
  nrow = 2
)
gammacap <- matrix(
  data = c(
    2.00, 1.00, 0.50,
    1.00, 1.25, 1.00,
    0.50, 1.00, 2.00
  ),
  nrow = 3
)

rvcov_mvn_chol(
  rcap = 5,
  sigmacap = sigmacap,
  gammacap = gammacap,
  n = 100
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn_generic}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn_generic}()},
\code{\link{l_mvn}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()},
\code{\link{theta_helper}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
