% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rvcov_wishart_of_vechsigmacap.R
\name{rvcov_wishart_of_vechsigmacap}
\alias{rvcov_wishart_of_vechsigmacap}
\title{Generate Sample Covariances
from the Wishart Distribution - Half-Vectorization
of the Covariance Matrix Input}
\usage{
rvcov_wishart_of_vechsigmacap(rcap, x, df, list = TRUE, vec = FALSE)
}
\arguments{
\item{rcap}{Positive integer.
\code{R} variates.}

\item{x}{Numeric vector.
Parameter.
Half-vectorization of the covariance matrix
\eqn{ \mathrm{vech} \left( \boldsymbol{\Sigma} \right)}.}

\item{df}{Positive integer.
Parameter.
Degrees of freedom.}

\item{list}{Logical.
If \code{list = TRUE},
returns a list where each element is a covariance matrix.
If \code{list = FALSE},
returns a matrix where each row corresponds to
the vectorization of the output matrix (\code{vec = TRUE}) or
the half-vectorization of the output matrix (\code{vec = FALSE})}

\item{vec}{Logical.
This is only evaluated when \code{list = FALSE}.
If \code{vec = TRUE},
returns the vectorization of the covariance matrix
for each \code{R} variate.
If \code{vec = FALSE},
returns the half-covariance matrix
for each \code{R} variate.}
}
\value{
A list (\code{list = TRUE}) or matrix (\code{list = FALSE}).
}
\description{
Generate Sample Covariances
from the Wishart Distribution - Half-Vectorization
of the Covariance Matrix Input
}
\section{Dependencies}{
\itemize{
\item \code{\link[=sym_of_vech]{sym_of_vech()}}
}
}

\examples{
rvcov_wishart_of_vechsigmacap(
  rcap = 5,
  x = c(1, 0.5, 1),
  df = 100,
  list = FALSE
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn_generic}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn_generic}()},
\code{\link{l_mvn}()},
\code{\link{mvn_theta_helper}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_mvn_chol}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
\keyword{random}
\keyword{wishart}
