% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-l_mvn_generic.R
\name{l_mvn_generic}
\alias{l_mvn_generic}
\title{Log of the Likelihood of the Multivariate Normal Distribution - Generic}
\usage{
l_mvn_generic(x, mu, sigmacap)
}
\arguments{
\item{x}{Numeric vector of length \code{k}.
The ith vector of observations.}

\item{mu}{Numeric vector of length \code{k}.
Mean vector.}

\item{sigmacap}{Numeric matrix of size \code{k} by \code{k}.
Covariance matrix.}
}
\value{
A vector.
}
\description{
Calculates the log of the likelihood function
of the multivariate normal distribution
for the ith observation.
}
\examples{
n <- 5
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

xcap <- as.data.frame(
  t(
    rmvn_chol(
      n = n,
      mu = mu,
      sigmacap = sigmacap
    )
  )
)

lapply(
  X = xcap,
  FUN = l_mvn_generic,
  mu = mu,
  sigmacap = sigmacap
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{grad_l_mvn_generic}()},
\code{\link{grad_l_mvn}()},
\code{\link{hess_l_mvn_generic}()},
\code{\link{hess_l_mvn}()},
\code{\link{l_mvn}()},
\code{\link{negl_mvn}()},
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()},
\code{\link{theta_helper}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
