% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmvn_chol.R
\name{rmvn_chol}
\alias{rmvn_chol}
\title{Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition}
\usage{
rmvn_chol(n, mu, sigmacap, varnames = NULL, data_frame = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{mu}{Numeric vector.
Parameter.
Mean vector
\eqn{\boldsymbol{\mu}}.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{varnames}{Character vector
Optional variable names.}

\item{data_frame}{Logical.
If \code{data_frame = TRUE},
returns a \code{data.frame}.
If \code{data_frame = FALSE},
returns a \code{matrix}.}
}
\value{
A matrix (\code{data_frame = FALSE}) or data.frame (\code{data_frame = TRUE}).
}
\description{
Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition
}
\examples{
mu <- c(0, 0)
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

rmvn_chol(
  n = 5,
  mu = mu,
  sigmacap = sigmacap
)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{rmeans_mvn_chol_of_theta}()},
\code{\link{rmeans_mvn_chol}()},
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()},
\code{\link{theta_helper}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
