% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rmvn_chol_of_vechsrhocap.R
\name{rmvn_chol_of_vechsrhocap}
\alias{rmvn_chol_of_vechsrhocap}
\title{Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Strict Half-Vectorization
of the Correlation Matrix Input}
\usage{
rmvn_chol_of_vechsrhocap(n, x, varnames = NULL, data_frame = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{x}{Numeric vector.
Parameter.
Strict half-vectorization
of the correlation matrix
\eqn{\mathrm{vechs} \left( \mathbf{P} \right)}.}

\item{varnames}{Character vector
Optional variable names.}

\item{data_frame}{Logical.
If \code{data_frame = TRUE},
returns a \code{data.frame}.
If \code{data_frame = FALSE},
returns a \code{matrix}.}
}
\value{
A matrix (\code{data_frame = FALSE}) or data.frame (\code{data_frame = TRUE}).
}
\description{
Generate Data
from the Multivariate Normal Distribution
Using the Cholesky Decomposition - Strict Half-Vectorization
of the Correlation Matrix Input
}
\section{Dependencies}{
\itemize{
\item \code{\link[=sym_of_vechs]{sym_of_vechs()}}
\item \code{\link[=vechs]{vechs()}} (test)
\item \code{\link[=vechsnames]{vechsnames()}} (test)
}
}

\examples{

x <- rmvn_chol_of_vechsrhocap(
  n = 100,
  x = 0.5
)

colMeans(x)
cov(x)
}
\seealso{
Other Multivariate Normal Distribution Functions: 
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution Functions}
\keyword{multiNorm}
