% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rvcov_wishart.R
\name{rvcov_wishart}
\alias{rvcov_wishart}
\title{Generate Sample Variance-Covariance Matrix from the Wishart Distribution}
\usage{
rvcov_wishart(rcap, sigmacap, df, vector = FALSE)
}
\arguments{
\item{rcap}{Positive integer.
\code{R} variates.}

\item{sigmacap}{Numeric matrix.
Parameter.
Covariance matrix
\eqn{\boldsymbol{\Sigma}}.}

\item{df}{Positive integer.
Parameter.
Degrees of freedom.}

\item{vector}{Logical.
If \code{vector = TRUE},
returns the half-vectorization of the covariance matrix
for each \code{R} variate.
If \code{vector = FALSE},
returns the covariance matrix
for each \code{R} variate.}
}
\value{
A list (\code{vector = FALSE}) or matrix (\code{vector = TRUE}).
}
\description{
Generate Sample Variance-Covariance Matrix from the Wishart Distribution
}
\examples{
sigmacap <- matrix(
  data = c(
    1, 0.5, 0.5, 1
  ),
  nrow = 2
)

x <- rvcov_wishart(
  rcap = 100,
  sigmacap = sigmacap,
  df = 100,
  vector = TRUE
)

colMeans(x)
}
\seealso{
Other Multivariate Normal Distribution: 
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart_of_vechsigmacap}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution}
\keyword{multiNorm}
