% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiNorm-rvcov_wishart_of_vechsigmacap.R
\name{rvcov_wishart_of_vechsigmacap}
\alias{rvcov_wishart_of_vechsigmacap}
\title{Generate Sample Covariances from the Wishart Distribution - vech of Sigma}
\usage{
rvcov_wishart_of_vechsigmacap(n, x, df, vector = FALSE)
}
\arguments{
\item{n}{Positive integer.
\code{n} variates.}

\item{x}{Numeric vector.
Half-vectorization of the covariance matrix.}

\item{df}{Positive integer.
Degrees of freesom.}

\item{vector}{Logical.
If \code{vector = TRUE},
returns the half-vectorization of the covariance matrix
for each \code{n} variate.
If \code{vector = FALSE},
returns the covariance matrix
for each \code{n} variate.}
}
\value{
A list (\code{vector = FALSE}) or matrix (\code{vector = TRUE}).
}
\description{
Generate Sample Covariances from the Wishart Distribution - vech of Sigma
}
\section{Dependencies}{
\itemize{
\item \code{\link[=sym_of_vech]{sym_of_vech()}}
\item \code{\link[=vech]{vech()}} (test)
\item \code{\link[=vechnames]{vechnames()}} (test)
}
}

\examples{
x <- rvcov_wishart_of_vechsigmacap(
  n = 100,
  x = c(1, 0.5, 1),
  df = 100,
  vector = TRUE
)

colMeans(x)
}
\seealso{
Other Multivariate Normal Distribution: 
\code{\link{rmvn_chol_of_rhocap}()},
\code{\link{rmvn_chol_of_theta}()},
\code{\link{rmvn_chol_of_vechsrhocap}()},
\code{\link{rmvn_chol}()},
\code{\link{rvcov_wishart}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Multivariate Normal Distribution}
\keyword{multiNorm}
